Merge development into master before v1.0 release

+++ code-experiments/src/suite_largescale.c
@@ -43,8 +43,8 @@ static coco_problem_t *coco_get_largescale_problem(const size_t function,
     problem = f_ellipsoid_permblockdiag_bbob_problem_allocate(function, dimension, instance, rseed,
         problem_id_template, problem_name_template);
   } else {
-    coco_error("coco_get_largescale_problem(): cannot retrieve problem f%lu instance %lu in %luD", function,
-        instance, dimension);
+    coco_error("coco_get_largescale_problem(): cannot retrieve problem f%lu instance %lu in %luD",
+    		(unsigned long) function, (unsigned long) instance, (unsigned long) dimension);
     return NULL; /* Never reached */
   }
